Refactor and organize `io::device` and `io::types`
