+++ src/io/device/mod.rs
@@ -1,0 +1,15 @@
+//! Provide Low-level Device Functionality
+
+mod container;
+mod deferred;
+mod errors;
+mod traits;
+mod types;
+mod wrapper;
+
+pub use container::*;
+pub use deferred::*;
+pub use errors::*;
+pub use traits::*;
+pub use types::*;
+pub use wrapper::*;
