+++ src/io/device/wrapper.rs
@@ -1,0 +1,4 @@
+pub trait DeviceWrapper {
+    fn is_input(&self) -> bool;
+    fn is_output(&self) -> bool;
+}
