+++ src/io/types/direction.rs
@@ -1,0 +1,23 @@
+use serde::{Deserialize, Serialize};
+use core::fmt::{Display, Formatter};
+
+/// Enum used to classify direction of data flow in relation to system.
+///
+/// Input objects generate data from the outside world;
+/// output objects accept data, and manipulate the outside.
+#[derive(Debug, Serialize, Deserialize, Clone, Copy, Default, PartialEq)]
+pub enum IODirection {
+    #[default]
+    Input,
+    Output,
+}
+
+impl Display for IODirection {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let name = match self {
+            IODirection::Input => "Input",
+            IODirection::Output => "Output",
+        };
+        write!(f, "{}", name)
+    }
+}
