+++ src/io/types/kind.rs
@@ -1,0 +1,51 @@
+use serde::{Deserialize, Serialize};
+use std::fmt::{Display, Formatter};
+
+/// Defines I/O type.
+/// Intended to allow differentiation and classification.
+/// More may be appended as time goes on.
+/// This might be expanded to multiple levels of organization.
+#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize, Default)]
+pub enum IOKind {
+    #[default]
+    Unassigned,
+    Light,
+    Pressure,
+    Proximity,
+    RotationVector,
+    RelativeHumidity,
+    Temperature,
+    Voltage,
+    Current,
+    Color,
+    TVOC,
+    VocIndex,
+    NoxIndex,
+    Flow,
+    EC,
+    PH,
+}
+
+impl Display for IOKind {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let name = match self {
+            IOKind::Unassigned => "Unassigned",
+            IOKind::Light => "Light",
+            IOKind::Pressure => "Pressure",
+            IOKind::Proximity => "Proximity",
+            IOKind::RotationVector => "Rotation Vector",
+            IOKind::RelativeHumidity => "Relative Humidity",
+            IOKind::Temperature => "Ambient Temperature",
+            IOKind::Voltage => "Voltage",
+            IOKind::Current => "Current",
+            IOKind::Color => "Color",
+            IOKind::TVOC => "TVOC",
+            IOKind::VocIndex => "Voc Index",
+            IOKind::NoxIndex => "Nox Index",
+            IOKind::Flow => "Flow (liquid)",
+            IOKind::EC => "Electrical Conductivity (EC)",
+            IOKind::PH => "pH",
+        };
+        write!(f, "{}", name)
+    }
+}
